requirements.txt
fastapi
fastapi-users
fastapi-users-db-sqlalchemy
python-dotenv
aiosqlite
sqlalchemy